Added Jpa database

+++ src/main/java/pl/futurecollars/invoicing/db/file/FileBasedDatabase.java
@@ -21,7 +21,7 @@ public class FileBasedDatabase implements Database {
     private final JsonService jsonService;
 
     @Override
-    public int save(Invoice invoice) {
+    public long save(Invoice invoice) {
         try {
             invoice.setId(idService.getNextIdAndIncrement());
             filesService.appendLineToFile(databasePath, jsonService.toJson(invoice));
@@ -32,7 +32,7 @@ public class FileBasedDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> getById(int id) {
+    public Optional<Invoice> getById(long id) {
         try {
             return filesService.readAllLines(databasePath).stream().filter(line -> containsId(line, id))
                 .map(line -> jsonService.toObject(line, Invoice.class)).findFirst();
@@ -52,7 +52,7 @@ public class FileBasedDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
 
         try {
 
@@ -73,7 +73,7 @@ public class FileBasedDatabase implements Database {
     }
 
     @Override
-    public Optional<Invoice> delete(int id) {
+    public Optional<Invoice> delete(long id) {
         try {
             List<String> allInvoices = filesService.readAllLines(databasePath);
             var invoiceListExpectDeleted =
@@ -88,7 +88,7 @@ public class FileBasedDatabase implements Database {
 
     }
 
-    private boolean containsId(String line, int id) {
+    private boolean containsId(String line, long id) {
         return line.contains("\"id\":" + id + ",");
     }
 }
